Port `interface` snippet
